Separate VideoAnalyzer into Video, VideoAnalyzer, and VideoProcessor.
